requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.44.0
django-azure-sql-backend==2.1.1.1
et-xmlfile==1.0.1
haversine==2.3.0
idna==2.10
jdcal==1.4.1
msal==1.8.0
oauthlib==3.1.0
openpyxl==3.0.6
Pillow==8.1.0
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.2
urllib3==1.26.2
